+++ homepage/homepage/app/docs/[framework]/[...slug]/page.tsx
@@ -73,7 +73,5 @@ export async function generateStaticParams() {
     }
   }
 
-  console.log(paths);
-
   return paths;
 }
