+++ homepage/homepage/lib/docNavigationItems.ts
@@ -34,24 +34,6 @@ export const docNavigationItems = [
         href: "/docs/sync-and-storage",
         done: 100,
       },
-      // {
-      //   name: "React",
-      //   href: "/docs/project-setup/react",
-      //   done: 80,
-      // },
-      // {
-      //   name: "React Native",
-      //   href: "/docs/project-setup/react-native",
-      // },
-      // {
-      //   name: "Next.js",
-      //   href: "/docs/project-setup/next",
-      // },
-      // {
-      //   name: "VueJS",
-      //   href: "/docs/project-setup/vue",
-      //   done: 80,
-      // },
       {
         name: "Node.JS / server workers",
         href: "/docs/project-setup/server-side",
